requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
PyYAML==5.4.1
requests==2.25.1
soupsieve==2.2.1
urllib3==1.26.5
